feat: add ContactBottom component
