+++ packages/apps/crm/src/screens/contact/ContactDetailsScreen.js
@@ -1,12 +1,12 @@
 import React, {useEffect} from 'react';
-import {ScrollView, StyleSheet, View} from 'react-native';
-import {Screen, HeaderContainer, CircleButton} from '@axelor/aos-mobile-ui';
+import {ScrollView} from 'react-native';
+import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
 } from '@axelor/aos-mobile-core';
-import {ContactBody, ContactHeader} from '../../components';
+import {ContactBody, ContactBottom, ContactHeader} from '../../components';
 import {fetchContactEventById} from '../../features/eventSlice';
 import {fetchPartner} from '../../features/partnerSlice';
 import {getContact} from '../../features/contactSlice';
@@ -52,52 +52,9 @@ const ContactDetailsScreen = ({navigation, route}) => {
       <ScrollView>
         <ContactBody navigation={navigation} />
       </ScrollView>
-      <View style={styles.bottomContainer}>
-        <CircleButton
-          iconName="pen"
-          onPress={() =>
-            navigation.navigate('ContactFormScreen', {
-              idContact: idContact,
-            })
-          }
-        />
-      </View>
+      <ContactBottom idContact={idContact} navigation={navigation} />
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  item: {
-    marginHorizontal: 20,
-    marginVertical: 7,
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  headerContainerChildren: {
-    flexDirection: 'row',
-    marginLeft: '5%',
-    alignItems: 'center',
-  },
-  headerInfo: {
-    flexDirection: 'column',
-    marginLeft: '7%',
-  },
-  container: {
-    alignItems: 'center',
-  },
-  textTitle: {
-    fontWeight: 'bold',
-  },
-  bottomContainer: {
-    width: '90%',
-    flexDirection: 'row-reverse',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginVertical: 2,
-    marginBottom: 25,
-  },
-});
 
 export default ContactDetailsScreen;
